move tests properly
