+++ efaar_benchmarking/benchmarking.py
@@ -5,7 +5,6 @@ from typing import Dict, List, Optional, Set, Tuple
 
 import numpy as np
 import pandas as pd
-import pytest
 from geomloss import SamplesLoss
 from joblib import Parallel, delayed
 from scipy.stats import hypergeom, ks_2samp
@@ -15,16 +14,6 @@ from sklearn.utils import Bunch
 from torch import from_numpy
 
 import efaar_benchmarking.constants as cst
-from efaar_benchmarking.benchmarking import (
-    AggregateBy,
-    AverageType,
-    BenchmarkConfig,
-    compound_gene_benchmark,
-    compute_metrics,
-    compute_similarities,
-    process_predictions,
-    sample_for_item,
-)
 
 
 class AverageType(Enum):
@@ -845,201 +834,3 @@ def compound_gene_benchmark(
     results["baseline_auc_roc"] = baseline["auc_roc"].values
 
     return results
-
-
-
-
-@pytest.fixture
-def sample_truth_data():
-    """Create sample ground truth data with known properties."""
-    return pd.DataFrame({
-        'treatment': ['compound1', 'compound1', 'compound2', 'compound2'],
-        'gene_symbol': ['gene1', 'gene2', 'gene1', 'gene2'],
-        'nM_value': [100, 5000, 15000, 500],  # Mix of active, gray zone, and inactive
-    })
-
-@pytest.fixture
-def sample_map_data():
-    """Create sample embedding data with known similarities."""
-    features = pd.DataFrame({
-        'feat1': [1.0, 0.0, 0.0, 1.0],
-        'feat2': [0.0, 1.0, 1.0, 0.0]
-    }, index=['compound1_id', 'compound2_id', 'gene1_id', 'gene2_id'])
-    
-    metadata = pd.DataFrame({
-        'perturbation': ['compound1', 'compound2', 'gene1', 'gene2'],
-        'concentration': ['1.0', '1.0', np.nan, np.nan],
-    }, index=['compound1_id', 'compound2_id', 'gene1_id', 'gene2_id'])
-    
-    return Bunch(features=features, metadata=metadata)
-
-def test_compute_similarities_basic(sample_map_data):
-    """Test basic similarity computation."""
-    truth = pd.DataFrame({
-        'treatment': ['compound1', 'compound2'],
-        'gene_symbol': ['gene1', 'gene2']
-    })
-    
-    sims = compute_similarities(truth, sample_map_data, 'perturbation')
-    
-    assert isinstance(sims, pd.DataFrame)
-    assert sims.shape == (2, 2)  # 2 compounds x 2 genes
-    # compound1 should be perfectly similar to gene1 (same features)
-    assert np.isclose(sims.loc[('compound1', '1.0'), 'gene1'], 1.0)
-
-def test_compute_similarities_randomized(sample_map_data):
-    """Test randomized similarity computation."""
-    truth = pd.DataFrame({
-        'treatment': ['compound1'],
-        'gene_symbol': ['gene1']
-    })
-    
-    sims = compute_similarities(truth, sample_map_data, 'perturbation', randomize=True)
-    
-    assert isinstance(sims, pd.DataFrame)
-    assert sims.shape == (1, 1)
-    assert 0 <= sims.iloc[0, 0] <= 1
-
-def test_sample_for_item():
-    """Test negative sampling logic."""
-    item_data = pd.DataFrame({
-        'gene_symbol': ['gene1', 'gene2', 'gene3'],
-        'nM_value': [100, 5000, 15000]
-    })
-    pool = {'gene1', 'gene2', 'gene3', 'gene4', 'gene5'}
-    
-    items, labels = sample_for_item(
-        item_data,
-        pool,
-        activity_threshold=1000,
-        inactivity_threshold=10000,
-        target_col='gene_symbol',
-        min_negatives=2,
-        random_seed=42
-    )
-    
-    assert len(items) > 0
-    assert len(labels) == len(items)
-    assert sum(labels) == 1  # Only gene1 should be positive
-    assert 'gene3' not in items  # Should be excluded as it's in gray zone
-
-def test_compute_metrics():
-    """Test metric computation with known values."""
-    scores = np.array([0.9, 0.8, 0.3, 0.2])
-    labels = np.array([1, 0, 0, 1])
-    
-    ap, auc = compute_metrics(scores, labels)
-    
-    assert 0 <= ap <= 1
-    assert 0 <= auc <= 1
-    # With these specific values, AP should be less than 0.75
-    assert ap < 0.75
-
-def test_full_benchmark_macro_compound(sample_truth_data, sample_map_data):
-    """Test full benchmark with macro averaging by compound."""
-    config = BenchmarkConfig(
-        average_type=AverageType.MACRO,
-        aggregate_by=AggregateBy.COMPOUND,
-        min_negatives=2,
-        random_seed=42
-    )
-    
-    results = compound_gene_benchmark(
-        map_data=sample_map_data,
-        activity_threshold=1000,
-        inactivity_threshold=10000,
-        truth_data=sample_truth_data,
-        config=config
-    )
-    
-    assert isinstance(results, pd.DataFrame)
-    assert "concentration" in results.columns
-    assert "average_precision" in results.columns
-    assert "auc_roc" in results.columns
-    assert "baseline_average_precision" in results.columns
-    assert "baseline_auc_roc" in results.columns
-    assert len(results) > 0
-    assert "max" in results["concentration"].values
-
-def test_full_benchmark_micro_gene(sample_truth_data, sample_map_data):
-    """Test full benchmark with micro averaging by gene."""
-    config = BenchmarkConfig(
-        average_type=AverageType.MICRO,
-        aggregate_by=AggregateBy.GENE,
-        min_negatives=2,
-        random_seed=42
-    )
-    
-    results = compound_gene_benchmark(
-        map_data=sample_map_data,
-        activity_threshold=1000,
-        inactivity_threshold=10000,
-        truth_data=sample_truth_data,
-        config=config
-    )
-    
-    assert isinstance(results, pd.DataFrame)
-    assert all(results["baseline_auc_roc"] == 0.5)  # Random baseline should have 0.5 AUC-ROC
-
-def test_benchmark_edge_cases(sample_map_data):
-    """Test benchmark behavior with edge cases."""
-    # Empty truth data
-    empty_truth = pd.DataFrame(columns=['treatment', 'gene_symbol', 'nM_value'])
-    with pytest.raises(ValueError):
-        compound_gene_benchmark(
-            map_data=sample_map_data,
-            truth_data=empty_truth
-        )
-    
-    # All inactive data
-    all_inactive = pd.DataFrame({
-        'treatment': ['compound1'],
-        'gene_symbol': ['gene1'],
-        'nM_value': [20000]  # Above inactivity threshold
-    })
-    results = compound_gene_benchmark(
-        map_data=sample_map_data,
-        truth_data=all_inactive
-    )
-    assert len(results) > 0
-    assert all(results["average_precision"] == 0.0)  # No positives should give 0 AP
-
-def test_process_predictions(sample_truth_data, sample_map_data):
-    """Test prediction processing logic."""
-    config = BenchmarkConfig()
-    similarities = compute_similarities(sample_truth_data, sample_map_data, 'perturbation')
-    
-    predictions = process_predictions(
-        sample_truth_data,
-        similarities,
-        config,
-        thresholds=(1000, 10000),
-        pert_col='perturbation'
-    )
-    
-    assert isinstance(predictions, dict)
-    assert "max" in predictions
-    assert all(isinstance(p, list) for p in predictions.values())
-    
-    # Check prediction format
-    for conc, preds in predictions.items():
-        if preds:
-            scores, labels = preds[0]
-            assert isinstance(scores, np.ndarray)
-            assert isinstance(labels, np.ndarray)
-            assert len(scores) == len(labels)
-            assert set(labels).issubset({0, 1})
-
-def test_config_validation():
-    """Test configuration validation."""
-    # Invalid average type
-    with pytest.raises(ValueError):
-        BenchmarkConfig(average_type="invalid")
-    
-    # Invalid aggregate by
-    with pytest.raises(ValueError):
-        BenchmarkConfig(aggregate_by="invalid")
-    
-    # Invalid min negatives
-    with pytest.raises(ValueError):
-        BenchmarkConfig(min_negatives=-1)
